+++ src/core/prelude/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.42.0 (2021-04-02)
+
+#### :rocket: New Feature
+
+* [Added a feature of optional patterns with `Date.format`](https://github.com/V4Fire/Core/issues/163)
+
 ## v3.36.0 (2021-03-16)
 
 #### :boom: Breaking Change